docs [README]: Remove updated license statement
The line about contacting me (the owner) for licensing enquiries or requests is not required, as I cannot really grant people permission to do anything that GPL v3 doesn't allow. This is because of the open-source nature of the project, and the fact that many people have already contributed to it.