chore: Bump release version to 1.7.2
By accident I also included a fix for pre-commit hook here. The black hook was failing as it was unable to import a module from another package that is uses called click. this was because black v22.1.0 was being used in our hooks. v22.3.0 fixes this, so I changed pre-commit hook to use that black version.